# Adds nvembed model to quantization algorithm