--- a/vptq/layers/__init__.py
+++ b/vptq/layers/__init__.py
@@ -4,3 +4,4 @@
 # --------------------------------------------------------------------------
 
 from vptq.layers.model_base import AutoModelForCausalLM as AutoModelForCausalLM
+from vptq.layers.model_base import AutoModelForST as AutoModelForST