Add support for quantizing the NV-Embed model using VPTQ. Benchmarked using MTEB to evaluate performance against the `bfloat16` version. Current benchmarks are based on a single task, further testing with more meaningful retrieval tasks and parameter tuning would be interesting to explore.

Benchmarks below are provided for context only and are not part of this PR.

Batch sizes were set to `16` for `VPTQ-NVEmbed` and `4` for `bfloat16`, as these were the maximum sizes that fit on a 24GB GPU.

#### Results on a specific task in (MTEB)

| Model                 | Main Score | Evaluation Time (s) |
|-----------------------|------------|-----------------------|
| VPTQ-NVEmbed-128-1024 | 0.33361    | 753.76               |
| NV-Embed bfloat16     | 0.38469    | 874.00               |