@YangWang92 2024-12-10
Thanks for your contribution! I will mergeit into the main branch and release a new version on PyPI.