@bndos 2024-12-11
> Could you please provide any additional suggestions for our VPTQ project? Are there any areas where we need to make improvements?

As far as the algorithm is concerned, I’m still getting familiar with it and the codebase. However, one suggestion I’d make is to use the Ruff formatter (`ruff format`). I noticed that `ruff` is configured in the `pyproject.toml`, but the formatter itself isn’t being used. If this were my repo, I’d probably run `ruff format` and set up a GitHub action and pre-commit hook for `ruff check` and `ruff format`. The main advantage is avoiding formatting artifacts in PRs. Just my 2 cents! :)